requirement.txt
appdirs==1.4.3
click==6.7
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.1
Flask-SocketIO==2.8.6
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
newrelic==2.82.0.62
packaging==16.8
pyparsing==2.2.0
python-engineio==1.3.1
python-socketio==1.7.4
six==1.10.0
Werkzeug==0.12.1
https://github.com/miguelgrinberg/Flask-SocketIO/archive/master.zip